04.tool_use_generic_configration/requirements.txt
langchain==0.1.12
langchain-openai==0.0.8
openai==1.14.0
python-dotenv==1.0.1
requests==2.31.0
wikipedia==1.4.0 